rename and deprecate computed properties
